+++ src/settings.h
@@ -167,6 +167,7 @@ public:
 private:
     void setFlag(const QString flag);
     QSet<QString> getKeys(CfgType type);
+    QStringList parseFlags();
 
     Settings *config;
     const QCommandLineParser *parser;